Rewrite Timestamp refactoring
Ran `git revert --no-commit af5de8a747d174642d582af19e8a8f8b50757dc3`
and made substantial edits before committing, including retaining some
changes that would have been reverted and rewriting some others.  This
compiles and most tests pass (HBase has problems, there seem to be
problems related to CString introduced in the last commit, I think a
recent BDB bugfix on titan05 is missing, maybe other failures).

This reverts commit af5de8a747d174642d582af19e8a8f8b50757dc3.

+++ titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/BackendTransaction.java
@@ -51,7 +51,7 @@ public class BackendTransaction implements TransactionHandle, LoggableTransactio
     private final KeyColumnValueStore indexStore;
 
     private final int maxReadRetryAttempts;
-    private final long retryStorageWaitTime;
+    private final int retryStorageWaitTime;
 
     private final Executor threadPool;
 
@@ -59,7 +59,7 @@ public class BackendTransaction implements TransactionHandle, LoggableTransactio
 
     public BackendTransaction(CacheTransaction storeTx, TransactionHandleConfig txConfig,
                               StoreFeatures features, KeyColumnValueStore edgeStore, KeyColumnValueStore indexStore,
-                              int maxReadRetryAttempts, long retryStorageWaitTime,
+                              int maxReadRetryAttempts, int retryStorageWaitTime,
                               Map<String, IndexTransaction> indexTx, Executor threadPool) {
         this.storeTx = storeTx;
         this.txConfig = txConfig;
